Hide menu on enter.
The assumption was made that when you hit enter, it will always be for navigational purposes and an item will be selected (causing the input to be cleared and the menu to hide). However, there is the use case that enter would be used for submitting a form (outside of this library). Then, if the input is cleared, the menu won't be hidden because we are listening for non-navigational keyups. I don't see any downside to always hide the menu when enter is pressed.